music_mode/requirements.txt
python-vlc
yt-dlp
ytmusicapi
SpeechRecognition
PyAudio
